small fixed in test comps
